rename(Amount): Payment -> Amount로 변경
